refactor: refine terrain logic, update GPU computations and parameters
- Adjusted terrain diffusity, transition factor, and forbidden parameter logic in pathfinding.
- Updated CUDA computations to use `double` for improved precision.
- Modified `LAND_TO_WATER` weight factor and kernel exponent for direction evaluation.
- Refined mixed GPU function for terrain walk steps and kernel pool data type conversion.
- Fixed inconsistencies in terrain transition matrix and reachability kernel application.

+++ src/cuda/mixed_gpu.h
@@ -32,7 +32,7 @@ typedef struct {
 } KernelPool;
 
 typedef struct {
-    float *kernel_pool;
+    double *kernel_pool;
     int kernel_pool_size;
 
     int *kernel_offsets;
